add keiko circular dependency correction ex3
