+++ src/exemple3/tribes.ts
@@ -1,9 +1,5 @@
 import { fred } from "./headOfTribe";
 
-export enum Tribes {
-  Kumo = "Kumo",
-}
-
 export const kumo = {
   headOfTribe: fred,
   technology: "serverless",
