+++ src/exemple3/types.ts
@@ -1,0 +1,11 @@
+export enum M33 {
+  Theodo = "Theodo",
+}
+
+export enum HeadOfTribeName {
+  Fred = "Fred",
+}
+
+export enum Tribes {
+  Kumo = "Kumo",
+}
